refactor: customer dto usando lombok
