+++ backend/customer/src/main/java/com/ms/customer/dto/AddressDTO.java
@@ -1,7 +1,15 @@
 package com.ms.customer.dto;
 
-public class AddressDTO {
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class AddressDTO {
     private String cep;
     private String uf;
     private String cidade;
@@ -9,74 +17,4 @@ public class AddressDTO {
     private String rua;
     private String numero;
     private String complemento;
-
-    public AddressDTO() {
-    }
-
-    public AddressDTO(String cep, String uf, String cidade, String bairro, String rua, String numero, String complemento) {
-        this.cep = cep;
-        this.uf = uf;
-        this.cidade = cidade;
-        this.bairro = bairro;
-        this.rua = rua;
-        this.numero = numero;
-        this.complemento = complemento;
-    }
-
-    // Getters e Setters
-    public String getCep() {
-        return cep;
-    }
-
-    public void setCep(String cep) {
-        this.cep = cep;
-    }
-
-    public String getUf() {
-        return uf;
-    }
-
-    public void setUf(String uf) {
-        this.uf = uf;
-    }
-
-    public String getCidade() {
-        return cidade;
-    }
-
-    public void setCidade(String cidade) {
-        this.cidade = cidade;
-    }
-
-    public String getBairro() {
-        return bairro;
-    }
-
-    public void setBairro(String bairro) {
-        this.bairro = bairro;
-    }
-
-    public String getRua() {
-        return rua;
-    }
-
-    public void setRua(String rua) {
-        this.rua = rua;
-    }
-
-    public String getNumero() {
-        return numero;
-    }
-
-    public void setNumero(String numero) {
-        this.numero = numero;
-    }
-
-    public String getComplemento() {
-        return complemento;
-    }
-
-    public void setComplemento(String complemento) {
-        this.complemento = complemento;
-    }
 }
