+++ backend/customer/src/main/java/com/ms/customer/dto/CheckMileResponseDTO.java
@@ -4,6 +4,15 @@ import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
 public class CheckMileResponseDTO {
 
     private Long codigo;
@@ -12,38 +21,4 @@ public class CheckMileResponseDTO {
     private Integer saldoMilhas;
 
     private List<TransitionDTO> transacoes;
-
-    public CheckMileResponseDTO() {
-    }
-
-    public CheckMileResponseDTO(Long codigo, Integer saldoMilhas, List<TransitionDTO> transacoes) {
-        this.codigo = codigo;
-        this.saldoMilhas = saldoMilhas;
-        this.transacoes = transacoes;
-    }
-
-    public Long getCodigo() {
-        return codigo;
-    }
-
-    public void setCodigo(Long codigo) {
-        this.codigo = codigo;
-    }
-
-    public Integer getSaldoMilhas() {
-        return saldoMilhas;
-    }
-
-    public void setSaldoMilhas(Integer saldoMilhas) {
-        this.saldoMilhas = saldoMilhas;
-    }
-
-    public List<TransitionDTO> getTransacoes() {
-        return transacoes;
-    }
-
-    public void setTransacoes(List<TransitionDTO> transacoes) {
-        this.transacoes = transacoes;
-    }
-
 }
