+++ backend/customer/src/main/java/com/ms/customer/dto/TransitionDTO.java
@@ -5,6 +5,15 @@ import java.time.LocalDateTime;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
 public class TransitionDTO {
 
     private LocalDateTime data;
@@ -21,64 +30,4 @@ public class TransitionDTO {
     private String codigoReserva;
 
     private String tipo;
-
-    public TransitionDTO() {
-    }
-
-    public TransitionDTO(LocalDateTime data, BigDecimal valorReais, Integer quantidadeMilhas, String descricao, String codigoReserva, String tipo) {
-        this.data = data;
-        this.valorReais = valorReais;
-        this.quantidadeMilhas = quantidadeMilhas;
-        this.descricao = descricao;
-        this.codigoReserva = codigoReserva;
-        this.tipo = tipo;
-    }
-
-    public LocalDateTime getData() {
-        return data;
-    }
-
-    public void setData(LocalDateTime data) {
-        this.data = data;
-    }
-
-    public BigDecimal getValorReais() {
-        return valorReais;
-    }
-
-    public void setValorReais(BigDecimal valorReais) {
-        this.valorReais = valorReais;
-    }
-
-    public Integer getQuantidadeMilhas() {
-        return quantidadeMilhas;
-    }
-
-    public void setQuantidadeMilhas(Integer quantidadeMilhas) {
-        this.quantidadeMilhas = quantidadeMilhas;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-
-    public void setDescricao(String descricao) {
-        this.descricao = descricao;
-    }
-
-    public String getCodigoReserva() {
-        return codigoReserva;
-    }
-
-    public void setCodigoReserva(String codigoReserva) {
-        this.codigoReserva = codigoReserva;
-    }
-
-    public String getTipo() {
-        return tipo;
-    }
-
-    public void setTipo(String tipo) {
-        this.tipo = tipo;
-    }
 }
