+++ backend/customer/src/main/java/com/ms/customer/dto/updateMiles/UpdateMilesRequestDTO.java
@@ -1,0 +1,16 @@
+package com.ms.customer.dto.updateMiles;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Setter
+public class UpdateMilesRequestDTO {
+
+    private Integer quantidade;
+
+}
