+++ backend/customer/src/main/java/com/ms/customer/service/CustomerService.java
@@ -10,8 +10,8 @@ import com.ms.customer.dto.AddressDTO;
 import com.ms.customer.dto.CheckMileResponseDTO;
 import com.ms.customer.dto.CustomerResponseDTO;
 import com.ms.customer.dto.TransitionDTO;
-import com.ms.customer.dto.UpdateMilesRequestDTO;
-import com.ms.customer.dto.UpdateMilesResponseDTO;
+import com.ms.customer.dto.updateMiles.UpdateMilesRequestDTO;
+import com.ms.customer.dto.updateMiles.UpdateMilesResponseDTO;
 import com.ms.customer.exception.CustomerNotFoundException;
 import com.ms.customer.model.Cliente;
 import com.ms.customer.model.HistoricoMilhas;
